+++ CHANGELOG.md
@@ -7,6 +7,8 @@
     attached to tracing spans. Remove the `tracing` feature: tracing support is
     now always included.
 
+-   BugFix: `ToDeviceRequest` returned by `shareRoomKey(..)` always had an `undefined` `id` field.
+
 # matrix-sdk-crypto-wasm v2.0.0
 
 -   Updated rust sdk version to revision [c2bb76029ae6d99c741727e0f87abcd734377016](https://github.com/matrix-org/matrix-rust-sdk/commit/c2bb76029ae6d99c741727e0f87abcd734377016), including: